Basic Early Output System
We'll want output right from boot, but before any of the real kernel systems are running,
there isn't a way to register drivers for stuff like UART's or displays. So the "early_output"
system is designed to be dead-simple and just require a single "putchar" callback be registered per
output source. This way we can always try to get some output from the very beginning of the init process.