FDT Memory Detection
Adds a system for a memory "detection" system to hand off memory to the "boot allocator".
This is done by having the detector call boot_free on all of the non-overlapping regions in physical memory
which are availible for allocations. Then that memory should be able to be allocated by using "boot_alloc"

So new boot-allocator designs could be added as long as they implement boot_alloc and boot_free, without knowing
anything about how memory is actually found (ACPI, multiboot, FDT, etc.)