Fixed "syscon" Driver
"syscon" is just a generic set of registers, but the old driver was hardcoded to
treat all syscon nodes as if they are running on a QEMU riscv virt board without
any extra extensions or devices. That's not ideal. What we really want to do is look
for "syscon-poweroff" and "syscon-reboot" nodes in the tree and use those to figure out
how to shutdown and poweroff the system.